% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{$.path_chain}
\alias{$.path_chain}
\title{Access path_chain object}
\usage{
\method{$}{path_chain}(node, child)
}
\arguments{
\item{node}{path_chain}

\item{child}{nested path_chain name}
}
\value{
path_chain or character, if path indicates leaf of structure tree
}
\description{
Access path_chain object
}
