% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{$.path_elem}
\alias{$.path_elem}
\title{Access path_elem object}
\usage{
\method{$}{path_elem}(node, child)
}
\arguments{
\item{node}{path_elem}

\item{child}{nested path_elem name}
}
\value{
path_elem or chaacter, if path indicates leaf of structure tree
}
\description{
Access path_elem object
}
