% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaries.R
\name{create_sample_dir}
\alias{create_sample_dir}
\title{Create sample directory}
\usage{
create_sample_dir(root = ".", name = "files", override = FALSE)
}
\arguments{
\item{root}{root directory, where the new sample folder will be created}

\item{name}{name for the new sample folder}

\item{override}{boolean: override folder if it already exists}
}
\description{
Creates sample nested directory to test and learn path.chain package
}
\examples{
create_sample_dir(name = "files", override = TRUE)
list.files("files", all.files = TRUE, recursive = TRUE, include.dirs = TRUE)
fs::dir_tree("files")
}
