% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaries.R
\name{create_temp_dir}
\alias{create_temp_dir}
\title{Create temporary diectory and return its name}
\usage{
create_temp_dir(
  ...,
  warn = FALSE,
  recursive = FALSE,
  fsep = .Platform$file.sep
)
}
\arguments{
\item{...}{arbitrary character objects}

\item{warn}{warn, if folder already exists}

\item{recursive}{ogical. Should elements of the path other than the last be created?
If true, like the Unix command mkdir -p}

\item{fsep}{the path separator to use}
}
\description{
Create temporary diectory and return its name
}
\examples{
# Simply create and return temporal directory
create_temp_dir()
# Create temp dir and return concatenated path
# Keep in mind, that 'files' and 'report_2020.xls' will not be created.
create_temp_dir("files", "report_2020.xls")
}
