% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaries.R
\name{file_path}
\alias{file_path}
\title{Construct path to file without doubled separators}
\usage{
file_path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors}

\item{fsep}{the path separator to use}
}
\value{
character file path
}
\description{
Construct path to file without doubled separators
}
\examples{
file.path("files/", "data/", "cars.RData")
file_path("files/", "data/", "cars.RData")
}
