% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full-path-chain.R
\name{full_path_chain}
\alias{full_path_chain}
\title{Full path chain}
\usage{
full_path_chain(path = ".", naming = basename)
}
\arguments{
\item{path}{root path}

\item{naming}{naming function}
}
\value{
list of lists and character objects
}
\description{
Full path chain
}
