% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{on_path_not_exists}
\alias{on_path_not_exists}
\title{Function called if path does not exists}
\usage{
on_path_not_exists(fun)
}
\arguments{
\item{fun}{a function, one-side formula or NULL; if missing,
eturns value of the path.chain.on.path.not.exists option}
}
\description{
Function called if path does not exists
}
\examples{
on_path_not_exists(print)
on_path_not_exists()
}
