% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain-package.R
\docType{package}
\name{path.chain-package}
\alias{path.chain-package}
\alias{path.chain}
\title{path.chain: Concise Structure for Chainable Paths}
\description{
The aim of this package is to provide tools,
which allow us to represent directory structure as
nested R objects. It can be easily saved as .yaml
files so that we can later load it and use in our project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krzjoa/path.chain}
  \item \url{https://krzjoa.github.io/path.chain/}
  \item Report bugs at \url{https://github.com/krzjoa/path.chain/issues}
}

}
\author{
Krzysztof Joachimiak
}
\keyword{package}
