% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{path_chain_config}
\alias{path_chain_config}
\title{Load path chain from configuration file}
\usage{
path_chain_config(config.section, root.key = "kRoot")
}
\description{
Load path chain from configuration file
}
