% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.R
\name{relative_path_chain}
\alias{relative_path_chain}
\title{Create chainable path}
\usage{
relative_path_chain(path)
}
\arguments{
\item{path}{Path}
}
\value{
path_elem object
}
\description{
Create chainable path
}
\examples{
chained.path <- relative_path_chain(".")
}
