% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaries.R
\name{temp_path}
\alias{temp_path}
\title{Construct path to file in a temporary directory}
\usage{
temp_path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{arbitrary character objects}

\item{fsep}{the path separator to use.}
}
\value{
a path
}
\description{
Construct path to file in a temporary directory
}
\details{
Be careful: if you call this function, it only creates a path for temporary file/dir.
All the rest has to be created on your own, e.g. calling \link[base]{dir.create} function.
}
\examples{
temp_path("files", "report.csv")
}
