% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.list}
\alias{as.list}
\alias{as.list.path_chain}
\title{Convert object of type `path_chain` to list}
\usage{
\method{as.list}{path_chain}(x, ..., root.name = "root.dir")
}
\arguments{
\item{x}{a path_chain object}

\item{...}{elipsis for API consistency, does nothing}

\item{root.name}{key for root directory; default: 'root.dir'}
}
\description{
Convert object of type `path_chain` to list
}
\examples{
tmp <- create_temp_dir("files")
create_sample_dir(tmp)
path.chain <- path_chain(tmp)
as.list(path.chain)
}
