% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{as_path_chain}
\alias{as_path_chain}
\title{Create chainable path}
\usage{
as_path_chain(config.section, root.key = "kRoot")
}
\arguments{
\item{path}{Path}
}
\value{
path_chain object
}
\description{
This function returns
}
\examples{
nested.list <- list(kRoot = "root", "file1.txt", list("subdir", "file2.csv"))
chainable.path <- as_path_chain(nested.list)
}
