% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{as_path_chain}
\alias{as_path_chain}
\alias{as_path_chain.character}
\alias{as_path_chain.list}
\title{Create chainable path}
\usage{
as_path_chain(path, ...)

\method{as_path_chain}{character}(path)

\method{as_path_chain}{list}(config.section, root.key = "kRoot")
}
\arguments{
\item{path}{Path}
}
\value{
path_chain object
}
\description{
This function returns
}
\examples{
chainable.path <- as_path_chain(".")
}
