% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{$.path_chain}
\alias{$.path_chain}
\title{Access path_chain object}
\usage{
\method{$}{path_chain}(node, child)
}
\arguments{
\item{node}{path_chain}

\item{child}{nested path_chain name}
}
\value{
path_chain or character, if path indicates leaf of structure tree
}
\description{
Access path_chain object
}
\examples{
#' If we want to create our chain manually, we have start from the leaves
level2.b <- path_link("fileA.RData")
level2.a <- path_link("fileB.RData")
level1   <- path_link("data", list(level2.a = level2.a , level2.b = level2.b))
root     <- path_link("files", list(level1))
# Print root path
root$.
# Print file path using chaining
root$data$level2.a
}
