% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{create_path_chain}
\alias{create_path_chain}
\title{Get directory structure and create path_chain object}
\usage{
create_path_chain(path)
}
\arguments{
\item{path}{Path}
}
\value{
path_chain object
}
\description{
This function returns
}
\examples{
# A
chainable.path <- as_path_chain(".")
}
