% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{create_path_chain}
\alias{create_path_chain}
\title{Get directory structure and create path_chain object}
\usage{
create_path_chain(path, naming = basename)
}
\arguments{
\item{path}{root of the directory structure}

\item{naming}{function, which defines naming convention}
}
\value{
path_chain object
}
\description{
This function returns
}
\examples{
create_sample_dir(name = "files", override = TRUE)
fs::dir_tree("files")
chainable.path <- create_path_chain("files")
chainable.path$data$persons.csv
# With customized naming convention
chainable.path <- create_path_chain("files", naming = naming_k)
chainable.path$kData$kPersons
}
