% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaries.R
\name{create_sample_dir}
\alias{create_sample_dir}
\title{Create sample directory}
\usage{
create_sample_dir(path = "files", override = FALSE)
}
\arguments{
\item{path}{path for the new sample folder}

\item{override}{boolean: override folder if it already exists}
}
\description{
Creates sample nested directory to test and learn path.chain package
}
\examples{
tmp <- create_temp_dir("files")
create_sample_dir(tmp, override = TRUE)
list.files(tmp, all.files = TRUE, recursive = TRUE, include.dirs = TRUE)
fs::dir_tree(tmp)
}
