% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaries.R
\name{create_temp_dir}
\alias{create_temp_dir}
\title{Create temporary diectory and return its name}
\usage{
create_temp_dir(..., warn = FALSE)
}
\arguments{
\item{...}{arbitrary character objects (optional)}

\item{warn}{warn, if folder already exists}
}
\description{
Create temporary diectory and return its name
}
\details{
Be careful: if you call this function, it only creates the root temporary file.
Even if using concatenation, the only object created in file system is tmp directory.
All the rest has to be created on your own, e.g. calling \link[base]{dir.create} function.
}
\examples{
# Simply create and return temporal directory
create_temp_dir()
# Create temp dir and return concatenated path
# Keep in mind, that 'files' and 'report_2020.xls' will not be created.
create_temp_dir("files", "report_2020.xls")
}
