% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full-path-chain.R
\name{full_path_chain}
\alias{full_path_chain}
\title{Full path chain}
\usage{
full_path_chain(path = ".", naming = basename)
}
\arguments{
\item{path}{root path}

\item{naming}{naming function}
}
\value{
list of lists and character objects
}
\description{
`full_path_chain` represents another approach to creating chainable paths
In contrast to `path_chain`, this functon creates just a list with nested list with full paths as a leaves.
}
\examples{
fs::dir_list()
chainable.path <- full_path_chain(".")
chainable.path
}
