% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full-path-chain.R
\name{full_path_chain}
\alias{full_path_chain}
\title{Full path chain}
\usage{
full_path_chain(path = ".", root.name = ".", naming = basename)
}
\arguments{
\item{path}{root path}

\item{root.name}{naming convention for root directory}

\item{naming}{naming function}
}
\value{
list of lists and character objects
}
\description{
`full_path_chain` represents another approach to creating chainable paths
In contrast to `path_chain`, this functon creates just a list with nested list with full paths as a leaves.
}
\examples{
tmp <- create_temp_dir("files")
create_sample_dir(tmp, override = TRUE)
fs::dir_tree(tmp)
chainable.path <- full_path_chain(tmp)
chainable.path
}
