% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{naming_k}
\alias{naming_k}
\title{Naming convention, which adds k prefix for each key, capitalizes and removes file extension}
\usage{
naming_k(path)
}
\arguments{
\item{path}{full path or its element}
}
\description{
Naming convention, which adds k prefix for each key, capitalizes and removes file extension
}
\examples{
naming_k("path/to/myfile.txt")
# Using with full_path_chain
create_sample_dir(name = "files", override = TRUE)
full.path.chain <- full_path_chain("files", naming = naming_k)
full.path.chain
create_sample_dir(name = "files")
# Using with path_chain / create_path_chain
path.chain <- create_path_chain("files", naming = naming_k)
path.chain \%>\%
  as.list()
unlink(root, recursive = TRUE)
}
