% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain-package.R
\docType{package}
\name{path.chain-package}
\alias{path.chain-package}
\alias{path.chain}
\title{Package path.chain}
\description{
The aim of this package is to provide tools,
which allow us to represent directory structure as
nested R objects. It can be easily saved as .yaml
files so that we can later load it and use in pur project.
}
\author{
Krzysztof Joachimiak
}
\keyword{package}
