% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{path_chain}
\alias{path_chain}
\title{Path element - directory or file name}
\usage{
path_chain(node = NULL, children = NULL)
}
\arguments{
\item{node}{Current node name; character}

\item{children}{list of children - path_chains}
}
\value{
path_chain object
}
\description{
A 'link' of path_chain object
}
\examples{
# If we want to create our chain manually, we have start from the leaves
level2.b <- path_chain("fileA.RData")
level2.a <- path_chain("fileB.RData")
level1 <- path_chain("data", list(level2.a = level2.a , level2.b = level2.b))
root <- path_chain("files", list(level1))
# Print root path
root$.
# Print file path using chaining
root$data$level2.a
}
