% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{path_children}
\alias{path_children}
\title{Get children nodes, i.e. all the suddiectories in the given directory}
\usage{
path_children(path.chain)
}
\arguments{
\item{path.chain}{object of `path_chain` class}
}
\value{
a list of `path_chain` objects
}
\description{
Get children nodes, i.e. all the suddiectories in the given directory
}
\examples{
path.chain <- create_path_chain("files", override = TRUE)
path_children(path.chain)
unlink("files", recursive = TRUE)
}
