% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{print}
\alias{print}
\alias{print.path_chain}
\title{Print path_chain object}
\usage{
\method{print}{path_chain}(x, ...)
}
\description{
Print path_chain object
}
\examples{
level2.b <- path_chain("fileA.RData")
level2.a <- path_chain("fileB.RData")
level1   <- path_chain("data", list(level2.a = level2.a , level2.b = level2.b))
root     <- path_chain("files", list(level1))
print(root)

create_sample_dir(name = "files", override = TRUE)
chainable.path <- create_path_chain("files")
print(chainable.path)
unlink("files", recursive = TRUE)
}
