% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-chain.R
\name{relative_path_chain}
\alias{relative_path_chain}
\alias{path_chain}
\title{Create chainable path}
\usage{
path_chain(path)
}
\arguments{
\item{path}{Path}
}
\value{
path_elem object
}
\description{
This function returns
}
\examples{
chained.path <- path_chain(".")
}
